@@ -1,84 +1,90 @@
 <template>
   <div class="friend-list container mt-5">
     <!-- 검색 기능 -->
-    <div class="mb-3">
+    <div class="friend-search mb-3">
       <input v-model="searchQuery" type="text" class="form-control" placeholder="친구 검색">
     </div>
 
     <!-- 친구 리스트 -->
     <ul class="list-group">
-      <li v-for="friend in filteredFriends" :key="friend.id" class="list-group-item list-group-item-action bg-info text-white d-flex justify-content-between align-items-center" @click="showProfile(friend)">
-        <img :src="friend.profilePicture" alt="프로필 이미지" class="profile-image mr-2">
-        <span>{{ friend.name }}</span>
+      <li v-for="friend in filteredFriends" :key="friend.id" class="list-group-item" @click="showProfile(friend)">
+        <div class="friend-item">
+          <img :src="friend.profilePicture" alt="프로필 이미지" class="profile-image mr-2">
+          <span>{{ friend.name }}</span>
+        </div>
+        <div class="buttons">
+          <button class="btn btn-primary btn-sm" @click.stop="startChat(friend)">채팅</button>
+        </div>
       </li>
     </ul>
-
-    <!-- 친구 프로필 모달 -->
-    <div v-if="selectedFriend" class="modal fade show" tabindex="-1" role="dialog" @click="resetSelectedFriend">
-      <div class="modal-background">
-        <div class="modal-dialog modal-dialog-centered" role="document">
-          <div class="modal-content">
-            <div class="modal-header">
-              <h5 class="modal-title">친구 프로필</h5>
-              <button type="button" class="close" data-dismiss="modal" @click="resetSelectedFriend">
-                <span>&times;</span>
-              </button>
-            </div>
-            <div class="modal-body">
-              <p><strong>이름:</strong> {{ selectedFriend.name }}</p>
-              <p><strong>나이:</strong> {{ selectedFriend.age }}</p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
   </div>
 </template>
 
-<script>
-export default {
-  data() {
-    return {
-      searchQuery: '',
-      friends: [
-        { id: 1, name: '김철수', age: 25 },
-        { id: 2, name: '김영희', age: 30 },
-        { id: 3, name: '김싸피', age: 28 },
-      ],
-      selectedFriend: null,
-    };
-  },
-  computed: {
-    filteredFriends() {
-      return this.friends.filter(friend =>
-        friend.name.toLowerCase().includes(this.searchQuery.toLowerCase())
-      );
-    },
-  },
-  methods: {
-    showProfile(friend) {
-      this.selectedFriend = friend;
-    },
-    resetSelectedFriend() {
-      this.selectedFriend = null;
-    },
-  },
+<script setup>
+import { ref, computed } from 'vue';
+import { useRouter } from 'vue-router';
+
+import FriendProfile from '@/components/Friend/FriendProfile.vue';
+import Chat from '@/components/Friend/Chat.vue'
+
+const router = useRouter();
+
+const showProfile = (friend) => {
+  // 모달 대신 페이지로 전환
+  const route = {
+    path: '/friend-profile',
+    component: FriendProfile,
+    props: { friend },
+  };
+
+  // Vue Router를 사용하여 페이지 전환
+  router.push(route);
+};
+
+const selectedFriend = ref(null);
+
+const goBackHandler = () => {
+  router.push('/friend')
+};
+
+const searchQuery = ref('');
+const friends = ref([
+  { id: 1, name: '김철수', age: 25 },
+  { id: 2, name: '김영희', age: 30 },
+  { id: 3, name: '김싸피', age: 28 },
+  { id: 4, name: '손흥민', age: 25 },
+  { id: 5, name: '봉준호', age: 30 },
+  { id: 6, name: '김준호', age: 28 },
+  { id: 7, name: '카리나', age: 25 },
+  { id: 8, name: '공유', age: 30 },
+  { id: 9, name: '황정민', age: 28 },
+]);
+
+const filteredFriends = computed(() => {
+  return friends.value.filter(friend =>
+    friend.name.toLowerCase().includes(searchQuery.value.toLowerCase())
+  );
+});
+
+const startChat = (friend) => {
+  selectedFriend.value = friend;
+  const route = {
+    path: '/chat',
+    component: Chat,
+  };
+  router.push(route);
 };
 </script>
 
-<style>
+<style scoped>
 .friend-list {
   max-width: 100%;
   margin: auto;
 }
 
-.list-group-item {
-  border-color: #87CEEB;
-}
-
 .modal-content {
-  background-color: #87CEEB;
-  color: #fff;
+  background-color: #EAF3F9;
+  color: black;
 }
 
 .profile-image {
@@ -97,7 +103,45 @@ export default {
 
 .modal-background {
   position: fixed;
-  top:0; left: 0; bottom: 0; right: 0;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
   background-color: rgba(0, 0, 0, 0.5);
+}
+
+.list-group-item {
+  border-color: #EAF3F9;
+  position: relative;
+  overflow: hidden;
+  background-color: #EAF3F9;
+  color: black;
+  margin-bottom: 10px;
+  border-radius: 24px;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.friend-item {
+  display: flex;
+  align-items: center;
+  position: relative;
+  z-index: 1;
+  padding-right: 80px;
+}
+
+.buttons {
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  display: flex;
+  align-items: center;
+  padding: 0 10px;
+  z-index: 2;
+}
+
+.buttons button {
+  margin-left: 5px;
 }
 </style>
