@@ -1,12 +1,12 @@
 <template>
-  <div id="app">
+  <div class="mt-5">
     <Sidebar />
 
     <!-- 하단 네비게이션 바 -->
     <div class="bottom-nav">
       <RouterLink to="/Main">Home</RouterLink>
       <RouterLink to="/Calendar">Calendar</RouterLink>
-      <RouterLink to="/Friend">Chat</RouterLink>
+      <RouterLink to="/Friend">Friends</RouterLink>
       <RouterLink to="/Meeting">Meeting</RouterLink>
       <RouterLink to="/mypage">MyPage</RouterLink>
     </div>
@@ -15,14 +15,15 @@
 </template>
 
 <script>
-import Sidebar from '@/views/Sidebar.vue'
+import Sidebar from '@/components/Sidebar.vue'
 import TodoList from '@/views/TodoList.vue'
 import Main from '@/views/Main.vue'
-import Calendar from '@/views/Calendar.vue'
-import Friend from '@/views/Friend.vue'
 import MyPage from '@/views/MyPage.vue'
 
 import { RouterLink, RouterView } from 'vue-router'
+
+let today = new Date()
+console.log(today)
 
 export default {
   name: 'App',
@@ -35,8 +36,6 @@ export default {
     Sidebar,
     TodoList,
     Main,
-    Calendar,
-    Friend,
     MyPage,
     RouterView,
     RouterLink,
@@ -46,19 +45,16 @@ export default {
       this.is_modal_valid = false
     }
   }
-  
-  // Vue 인스턴스의 데이터와 메소드를 여기에 정의합니다.
 }
 </script>
 
 <style>
 /* 전체 앱 스타일링 */
 #app {
-  font-family: 'Avenir', Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
+    font-family: 'Avenir', Helvetica, Arial, sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    color: #2c3e50;
 }
 
 /* 상단 바 스타일링 */
@@ -76,25 +72,6 @@ export default {
 
 
 /* 할 일 목록 스타일링 */
-.todo-section {
-  background-color: #ecf0f1;
-  padding: 10px;
-  margin: 10px 0;
-}
-
-.todo-date {
-  font-weight: bold;
-  margin-bottom: 5px;
-}
-
-.todo-item {
-  display: flex;
-  align-items: center;
-}
-
-.todo-item input[type="checkbox"] {
-  margin-right: 5px;
-}
 
 /* 하단 네비게이션 바 스타일링 */
 .bottom-nav {
@@ -106,7 +83,9 @@ export default {
   display: flex;
   justify-content: space-around;
   background-color: #f3f3f3;
-  padding: 10px 0;
+  padding: 10px;
+  z-index: 999;
+  height: 60px
 }
 
 
