@@ -1,19 +1,41 @@
+// router/index.js
 import { createRouter, createWebHistory } from 'vue-router'
 import Main from '@/views/Main.vue'
+import Goal from '@/views/GoalView.vue'
+import GoalList from '@/views/GoalList.vue'
 import Board from '@/views/Board.vue'
 import Calendar from '@/views/Calendar.vue'
 import Friend from '@/views/Friend.vue'
+import Chat from '@/components/Friend/Chat.vue'
 import Meeting from '@/views/Meeting.vue'
+import Habit from '@/views/HabitView.vue'
 import MyPage from '@/views/MyPage.vue'
-import Voice from '@/components/Voice.vue'
+import Voice from '@/views/Voice/Voice.vue'
+import VoiceTrainer from '@/views/Voice/VoiceTrainer.vue'
+import Notification from '@/components/Notifications.vue'
+import subscribe from '@/components/subscribe_page.vue'
+import pushnotify from '@/components/pushnotify_page.vue'
+import TodoList from '@/views/TodoList.vue'
+import CalendarDetail from '@/views/CalendarDetail.vue'
+import FriendProfile from '@/components/Friend/FriendProfile.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
 	{
+		name: '',
+		path: '/',
+		component: Main,
+	},
+	{
 		name: 'main',
 		path: '/main',
 		component: Main,
+	},
+	{
+		name: 'goal',
+		path: '/goal',
+		component: Goal,
 	},
     {
 		name: 'board',
@@ -30,22 +52,62 @@ const router = createRouter({
 		path: '/friend',
 		component: Friend,
 	},
+	{
+		name: 'chat',
+		path: '/chat',
+		component: Chat,
+	},
     {
 		name: 'meeting',
 		path: '/meeting',
 		component: Meeting,
 	},
 	{
+		name: 'habit',
+		path: '/habit',
+		component: Habit,
+	},
+	{
 		name: 'mypage',
 		path: '/mypage',
 		component: MyPage,
 	},
+	{
+		path: '/voice',
+		name: 'Voice',
+		component: Voice
+	},
+	{
+		path: '/voice/trainer',
+		name: 'VoiceTrainer',
+		component: VoiceTrainer
+	},
+	{
+		path: '/subscribe',
+		name: 'subscribe',
+		component: subscribe
+	},
+	{
+		path: '/pushnotify',
+		name: 'pushnotify',
+		component: pushnotify
+	},
   {
-    name: 'voice',
-    path: '/voice',
-    component: Voice
+    path: '/calendar/calendarDetail', // 추후 바꿀 예정
+    name: 'CalendarDetail',
+    component: CalendarDetail
+  },
+  {
+    path: '/goallist',
+    name: 'GoalList',
+    component: GoalList
+  },
+  {
+	path: '/friend-profile',
+	component: FriendProfile
   }
   ]
 })
+
 
 export default router
