@@ -5,7 +5,7 @@
         <button class="navbar-toggler" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasNavbar" aria-controls="offcanvasNavbar" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon"></span>
         </button>
-        <router-link to="/" class="navbar-brand mx-auto" @click="closeOffcanvas">Todak Todak</router-link>
+        <a class="navbar-brand mx-auto" href="#">토닥토닥</a>
         <button type="button" class="btn btn-primary" data-bs-toggle="offcanvas" data-bs-target="#offcanvasNotifications" aria-controls="offcanvasNotifications">
           알림
         </button>
@@ -15,13 +15,29 @@
             <button type="button" class="btn-close" data-bs-dismiss="offcanvas" aria-label="Close"></button>
           </div>
           <div class="offcanvas-body">
-            <router-link to="/goal" @click="closeOffcanvas" class="nav-link">목표 만들기</router-link>
-            <router-link to="/calendar" @click="closeOffcanvas" class="nav-link">캘린더</router-link>
-            <router-link to="/meeting" @click="closeOffcanvas" class="nav-link">모임</router-link>
-            <router-link to="/board" @click="closeOffcanvas" class="nav-link">게시판</router-link>
-            <router-link to="/Friend" @click="closeOffcanvas" class="nav-link">친구</router-link>
-            <router-link to="/habit" @click="closeOffcanvas" class="nav-link">습관</router-link>
-            <router-link to="/mypage" @click="closeOffcanvas" class="nav-link">마이페이지</router-link>
+            <ul class="navbar-nav justify-content-end flex-grow-1 pe-3">
+              <li class="nav-item">
+                <RouterLink @click.native="closeSidebar" to="/calendar">캘린더</RouterLink>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="#">캘린더</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="#">모임</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="#">게시판</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="#">친구</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="#">습관</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link" href="#">마이페이지</a>
+              </li>
+            </ul>
           </div>
         </div>
       </div>
@@ -30,20 +46,24 @@
 </template>
 
 <script>
-import Friend from '@/views/Friend.vue'
+import Calendar from '@/views/Calendar.vue'
+
+import { RouterLink, RouterView } from 'vue-router'
 
 export default {
   name: 'Sidebar',
   components: {
-    Friend,
+    Calendar,
+    RouterView,
+    RouterLink,
   },
   methods: {
-    closeOffcanvas() {
-      // 오프캔버스 닫기
-      const offcanvasNavbar = new bootstrap.Offcanvas(document.getElementById('offcanvasNavbar'));
-      offcanvasNavbar.hide();
+      // 변경된 부분: 페이지 이동 시 Sidebar를 닫는 메서드
+      closeSidebar() {
+        console.log('closeSidebar method called');
+        this.$root.$emit('bv::toggle::offcanvas', 'offcanvasNavbar');
+      },
     },
-  },
 }
 </script>
 
